requirements.txt
fastapi
uvicorn
PyMuPDF
gradio
requests
pytest
python-magic